UP-2196: Get blockchain explorer settings from global web-ui project settings into verification widget
