+++ .gitignore
@@ -43,3 +43,4 @@ UserInterfaceState.xcuserstate
 /*-import.json
 /preprocess-replace-secrets.sh
 package-lock.json
+/src/assets/constants/
