+++ src/app/models/iubirch-blockchain-net.ts
@@ -1,0 +1,3 @@
+export interface IUbirchBlockchainNet {
+  url: string;
+}
