+++ src/app/models/iubirch-blockchain.ts
@@ -1,0 +1,6 @@
+import {IUbirchBlockhainTransidCheckUrl} from './iubirch-blockhain-transid-check-url';
+
+export interface IUbirchBlockchain {
+  nodeIcon: string;
+  explorerUrl: IUbirchBlockhainTransidCheckUrl;
+}
