+++ src/app/models/iubirch-blockhain-settings.ts
@@ -1,0 +1,5 @@
+import {IUbirchBlockchain} from './iubirch-blockchain';
+
+export interface IUbirchBlockhainSettings {
+  [key: string]: IUbirchBlockchain;
+}
