+++ src/app/models/iubirch-blockhain-transid-check-url-map.ts
@@ -1,0 +1,5 @@
+import {IUbirchBlockhainTransidCheckUrl} from './iubirch-blockhain-transid-check-url';
+
+export interface IUbirchBlockhainTransidCheckUrlMap {
+  [key: string]: IUbirchBlockhainTransidCheckUrl;
+}
