+++ src/app/models/iubirch-blockhain-transid-check-url.ts
@@ -1,0 +1,7 @@
+import {IUbirchBlockchainNet} from './iubirch-blockchain-net';
+
+export interface IUbirchBlockhainTransidCheckUrl {
+  bdr?: IUbirchBlockchainNet;
+  testnet?: IUbirchBlockchainNet;
+  mainnet?: IUbirchBlockchainNet;
+}
